Initial commit for Carthage support
