+++ .gitignore
@@ -2,6 +2,8 @@
 #
 # gitignore contributors: remember to update Global/Xcode.gitignore, Objective-C.gitignore & Swift.gitignore
 
+.gitmodules
+
 ## Build generated
 build/
 DerivedData
@@ -38,9 +40,10 @@ Pods/
 # Carthage
 #
 # Add this line if you want to avoid checking in source code from Carthage dependencies.
-# Carthage/Checkouts
-
-Carthage/Build
+#Carthage/Checkouts
+#Carthage/Build
+Carthage/
+Cartfile.resolved
 
 # fastlane
 #
