+++ .travis.yml
@@ -2,4 +2,4 @@ language: objective-c
 osx_image: xcode7.3
 podfile: Demo/iOS/Podfile
 #script: xctool -workspace Demo/iOS/WKZombieDemo.xcworkspace -scheme WKZombieDemo -sdk iphonesimulator build #test
-script: xcodebuild -workspace Demo/iOS/WKZombieDemo.xcworkspace -scheme WKZombieDemo -sdk iphonesimulator build test
+script: xcodebuild -workspace WKZombie.xcworkspace -scheme WKZombie -sdk iphonesimulator build test
