--- Demo/OSX/Podfile
@@ -1,9 +1,0 @@
-platform :osx, '10.11'
-
-source 'https://github.com/CocoaPods/Specs.git'
-
-inhibit_all_warnings!
-use_frameworks!
-link_with "WKZombieDemo", "Tests"
-
-pod 'WKZombie', :path => '../../'
