--- Demo/iOS/Podfile
@@ -1,10 +1,0 @@
-platform :ios, '9.1'
-
-source 'https://github.com/CocoaPods/Specs.git'
-
-inhibit_all_warnings!
-use_frameworks!
-
-link_with "WKZombieDemo", "Tests"
-
-pod 'WKZombie', :path => '../../'
