+++ Example/Example OSX/Assets.xcassets/AppIcon.appiconset/Contents.json
