+++ Example/Example iOS/AppDelegate.swift
