+++ Example/Example iOS/Assets.xcassets/AppIcon.appiconset/Contents.json
