+++ Example/Example-Bridging-Header.h
@@ -1,0 +1,14 @@
+//
+//  Example-Bridging-Header.h
+//  Example
+//
+//  Created by Mathias Köhnke on 07/04/16.
+//  Copyright © 2016 Mathias Köhnke. All rights reserved.
+//
+
+#ifndef Example_Bridging_Header_h
+#define Example_Bridging_Header_h
+
+#import <hpple/TFHpple.h>
+
+#endif /* Example_Bridging_Header_h */
