+++ WKZombie.podspec
@@ -25,5 +25,5 @@ Pod::Spec.new do |s|
 
   s.requires_arc = true
 
-  s.dependency 'hpple', '0.2.0'
+  s.dependency 'hpple', '0.2.0' 
 end
