+++ WKZombie.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "group:Carthage/Checkouts/hpple/Framework/Hpple.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "container:WKZombie.xcodeproj">
+   </FileRef>
+</Workspace>
